# docs/development/api/openapi.json
{
  "components": {
    "schemas": {
      "ErrorBody": {
        "description": "Standard error payload returned by every failed request.\n\nThe envelope is uniform across every error the server emits, including\nrouting fallbacks and extractor rejections (`API-039`).",
        "properties": {
          "_links": {
            "$ref": "#/components/schemas/SelfLinks",
            "description": "Link to the request that produced the error."
          },
          "error": {
            "description": "Human-readable description of the error.",
            "example": "An error message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "_links"
        ],
        "type": "object"
      },
...
      "GetMeResponse": {
        "description": "Current user returned by a successful lookup.",
        "properties": {
          "email": {
            "description": "Email address of the caller, when one was provided.",
            "format": "email",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/i64",
            "description": "Unique identifier of the caller."
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "Role of the caller."
          },
          "username": {
            "description": "Username of the caller.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "username"
        ],
        "type": "object"
      },
...
      "Link": {
        "description": "A single HAL link.",
        "properties": {
          "href": {
            "description": "Target URI reference of the link.",
            "type": "string"
          },
          "templated": {
            "description": "Whether `href` is a URI template, per RFC 6570.",
            "example": true,
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "required": [
          "href"
        ],
        "type": "object"
      },
...
      "Role": {
        "description": "Role of a `User`.",
        "enum": [
          "ADMIN",
          "STANDARD"
        ],
        "type": "string"
      },
      "SelfLinks": {
        "description": "The `self` link of a HAL representation.\n\nEvery resource representation carries the URI it was served from, so clients\nnever construct or hardcode it themselves (`API-032`, `API-037`).",
        "properties": {
          "self": {
            "$ref": "#/components/schemas/Link",
            "description": "Link to the resource itself."
          }
        },
        "required": [
          "self"
        ],
        "type": "object"
      },
...
      "i64": {
        "format": "int64",
        "type": "integer"
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP API of the Mnemorium service",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Mnemorium API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/user/me": {
      "get": {
        "description": "Returns the caller's own profile — identifier, username, email and role.",
        "operationId": "get_me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMeResponse"
                }
              }
            },
            "description": "Current user"
          },
          "401": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "500": {
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "summary": "Fetch the current user",
        "tags": [
          "user"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://0.0.0.0:4080/api/v1"
    }
  ],
...
}